♻️ refactor directive
