--- packages/directive-headers/.gitignore
@@ -1,8 +1,0 @@
-# Logs
-logs
-*.log
-npm-debug.log*
-
-node_modules
-_build
-.DS_Store
