--- packages/directive-headers/tsconfig-esm.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "esnext",
-    "outDir": "./_build/esm/"
-  }
-}
