--- packages/directive-no-auth/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "directive-no-auth",
-  "version": "1.0.0",
-  "type": "module",
-  "main": "_build/cjs/index.js",
-  "module": "_build/esm/index.js",
-  "types": "_build/esm/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./_build/esm/index.js",
-      "require": "./_build/cjs/index.js"
-    }
-  },
-  "files": [
-    "_build/**/*"
-  ],
-  "scripts": {
-    "build:cjs": "tsc --project tsconfig-cjs.json",
-    "build:esm": "tsc --project tsconfig-esm.json",
-    "build": "rm -rf _build && npm run build:esm && npm run build:cjs && node ./scripts/prepare-package-json",
-    "pack": "npm run build && npm pack --pack-destination ../graphql-mesh/local-pkg"
-  },
-  "devDependencies": {
-    "@types/node": "^20.11.19",
-    "tslib": "^2.6.2",
-    "typescript": "^5.2.2"
-  },
-  "peerDependencies": {
-    "@graphql-mesh/cache-localforage": "*",
-    "@graphql-tools/utils": "*",
-    "@graphql-mesh/types": "*",
-    "@graphql-mesh/utils": "*",
-    "graphql": "*"
-  }
-}
