--- packages/directive-no-auth/tsconfig-cjs.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "./_build/cjs/"
-  }
-}
