--- packages/directive-no-auth/tsconfig.json
@@ -1,30 +1,0 @@
-{
-  "exclude": ["node_modules", "**/*.test.ts"],
-  "include": ["src/**/*"],
-  "compilerOptions": {
-    "allowJs": false,
-    "allowSyntheticDefaultImports": true,
-    "baseUrl": ".",
-    "declaration": true,
-    "experimentalDecorators": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "importHelpers": true,
-    "isolatedModules": false,
-    "lib": ["ESNext"],
-    "moduleResolution": "node",
-    "noEmit": false,
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "noUnusedParameters": true,
-    "paths": {
-      "src/*": ["src/*"],
-      "typings": ["src/typings.ts"]
-    },
-    "resolveJsonModule": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "target": "es2020"
-  }
-}
