+++ packages/graphql-mesh/directives/no-auth.ts
@@ -2,34 +2,26 @@ import { defaultFieldResolver, GraphQLSchema } from 'graphql'
 import { MeshTransform } from '@graphql-mesh/types'
 import { MapperKind, mapSchema } from '@graphql-tools/utils'
 
-export default class HeadersDirectiveTransform implements MeshTransform {
+export default class NoAuthDirectiveTransform implements MeshTransform {
   noWrap = true
 
   transformSchema(schema: GraphQLSchema) {
     return mapSchema(schema, {
       [MapperKind.OBJECT_FIELD]: (fieldConfig) => {
-        const originalResolver =
-          fieldConfig.resolve != null ? fieldConfig.resolve : defaultFieldResolver
+        const originalResolver = fieldConfig.resolve ?? defaultFieldResolver
 
         const resolver = async (next: any, _source: any, _args: any, context: any, info: any) => {
           const { directives } = info.fieldNodes[0]
-          const headersDirective = directives.find(
-            (directive: { name: { value: string } }) => directive.name.value === 'headers'
+          const noAuthDirective = directives.find(
+            (directive: { name: { value: string } }) => directive.name.value === 'noAuth'
           )
 
-          if (headersDirective) {
-            const { value } = headersDirective.arguments[0]
-
-            value?.values?.forEach((item: { fields: [any, any] }) => {
-              const [headerName, headerValue] = item.fields
-              context = {
-                ...context,
-                headers: {
-                  ...context.headers,
-                  [headerName.value.value.toLowerCase()]: headerValue.value.value
-                }
-              }
-            })
+          /**
+           * In order to set headers for the request, we need override authorization headers
+           * an pass it to execute function
+           */
+          if (noAuthDirective) {
+            context = { ...context, headers: { ...context.headers, authorization: '' } }
           }
 
           return await next(context)
@@ -57,14 +49,9 @@ export default class HeadersDirectiveTransform implements MeshTransform {
   }
 }
 
-export const headersDirectiveTypeDef: string = /* GraphQL */ `
-  input Header {
-    key: String
-    value: String
-  }
-
+export const noAuthDirectiveTypeDef: string = /* GraphQL */ `
   """
-  This directive is used to add headers to the request.
+  This directive is used to disable the authorization header for the request
   """
-  directive @headers(input: [Header]) on FIELD
+  directive @noAuth on FIELD
 `
